+++ src/vm/vm_trace.cpp
@@ -5,12 +5,15 @@
 
 #include "vm/vm_internal.hpp"
 
+#include "common/features.hpp"
 #include "core/function.hpp"
 #include "core/gc_string.hpp"
 #include "debug/trace_sink.hpp"
 #include "debug/trace_types.hpp"
 #include "debug/value_serializer.hpp"
+#if LUA_CPP_ENABLE_DEBUGGER
 #include "debugger/debug_runtime.hpp"
+#endif
 #include "vm/state/call_info.hpp"
 #include "vm/state/lua_state.hpp"
 #include "vm/state/stack.hpp"
@@ -133,22 +136,26 @@ bool isTraceDiffEnabled(RuntimeServices& services) {
 namespace VM::detail {
 
 void dispatchCallHook(LuaState* L) {
+#if LUA_CPP_ENABLE_DEBUGGER
     Debugger::DebugController* debugger = L->getGlobalState().getDebugController();
     if (debugger != nullptr && debugger->semanticSafepoint(*L, Debugger::DebugSemanticEvent::FunctionEnter) ==
                                    Debugger::DebugSafepointResult::TerminateExecution) [[unlikely]] {
         throw RuntimeError("debugger requested execution termination");
     }
+#endif
     if (L->hasDebugHookMask(HookMaskCall)) {
         L->callDebugHook(DebugHookEvent::Call);
     }
 }
 
 void dispatchReturnHook(LuaState* L) {
+#if LUA_CPP_ENABLE_DEBUGGER
     Debugger::DebugController* debugger = L->getGlobalState().getDebugController();
     if (debugger != nullptr && debugger->semanticSafepoint(*L, Debugger::DebugSemanticEvent::FunctionReturn) ==
                                    Debugger::DebugSafepointResult::TerminateExecution) [[unlikely]] {
         throw RuntimeError("debugger requested execution termination");
     }
+#endif
     if (L->hasDebugHookMask(HookMaskReturn)) {
         i32 tailcalls = 0;
         if (L->getCurrentCI() < L->getCallStack().size()) {