+++ tests/lua/debugger/basic_breakpoint.lua
@@ -3,4 +3,3 @@ local count = 2
 local result = greeting .. " " .. count
 
 return result
-